Extracting the user context for SingUp
Changed from having the context in signup.jsx to an independent component UserContext.jsx
	modified:   src/App.jsx
	modified:   src/components/PreRegister.jsx
	modified:   src/components/Register.jsx
	modified:   src/components/Register2.jsx
	modified:   src/components/RegisterVehicle.jsx
	modified:   src/pages/Singup.jsx
	new file:   src/services/UserContext.jsx
