+++ src/App.jsx
@@ -7,6 +7,7 @@ import { ProtectedRoute } from "./services/ProtectedRoutes.jsx";
 import { AuthContextProvider } from "./services/AuthContext";
 import {HomeBidder} from "./pages/HomeBidder";
 import {HomeRider} from "./pages/HomeRider";
+import { UserContextProvider } from "./services/UserContext";
 
 function App() {
 	return (
@@ -41,7 +42,9 @@ function App() {
 					path="/bidder"
 					element={
 						<ProtectedRoute accessBy="authenticated">
-							<HomeBidder />
+							<UserContextProvider>
+								<HomeBidder />
+							</UserContextProvider>							
 						</ProtectedRoute>
 					}
 				></Route>
@@ -49,7 +52,9 @@ function App() {
 					path="/SingUp"
 					element={
 						<ProtectedRoute accessBy="non-authenticated">
-							<SingUp />								
+							<UserContextProvider>
+								<SingUp />
+							</UserContextProvider>							
 						</ProtectedRoute>
 					}>
 				</Route>
